requirements.txt
azure-mgmt-datafactory>=2.0.0
azure-identity>=1.12.0
databricks-sdk>=0.12.0
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
slack-sdk>=3.21.3
tenacity>=8.2.2
python-dateutil>=2.8.2
structlog>=23.1.0 